requirements.txt
flask
azure-cosmos
msal
azure-identity
azure-mgmt-authorization
